Create Day_2.Jewels and Stones
This problem is quite simple and I solved it in just 1 and half a minute. I didn't expect it to be so simple. I am very glad that I am achieving this like on youtube, programmer solves the problem with their timer on. Let it be like that. I wish myself godspeed and have this excitement continue so that my zeal for programming live on.