Create Day_3.Ransom Note
This problem solution is a bit lengthy but it is the only solution came to my mind then. Anyway, later I found a better solution. Okay, it is part of the journey,' What I learn today, will apply tomorrow'.At least, these are my honest approach and I am feeling good about it.